Add Couchbase vars to the top of the .sample.env.
One of the first steps in the README.md is to copy the sample to .env
and edit it.  That's rather confusing as there's nothing Couchbase
related in there, but there are others such as MySQL.

Based on code inspection, it looks like this is what would be needed.
I left it at what the defaults are underneath so a user can edit.